Last Commit

+++ database/migrations/2022_09_22_151532_create_exams_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
         Schema::create('exams', function (Blueprint $table) {
             $table->id();
             $table->integer('user_id');
-            $table->integer('total_marks');
+            $table->float('total_marks');
             $table->timestamps();
         });
     }
